automatically manage passphrase and key on mac with bash script
Using a passphrase is more secure, and on mac adding it to the keychain maintains the
convenience when using the key.
This acheives two things:
1. a passphrease is automatically generated
2. key is automatically added with passphrase (and old one removed) to
the keychain

As consequence:
3. no build up of keys in key agent
4. security at rest *and* convenience (remote login requires no extra
   password if one has logged in his computer)

This commit fixes the bash script, the python script could be fixed analogously.
If the service would accept an optional passphrase we could avoid storing a private
key without passphrase at any point (always secure at rest).